Move DB fn into db_context.go
